bindgen: Only generate bindings for `I?ADLX*` items
These headers include `Windows.h`, resulting in a 335k(!!) bindings
file that is very hard for `rustc` (and `rust-analyzer`) to process.
In addition `cargo` seems to fail generating a proper fingerprint for
it, resulting in repeated re-runs of `build.rs` that take upwards of 2
minutes each.

Rather than declaring every API that we _do not want_, list only the
regex-based API for items that we _do want_ to generate bindings for.