interface: Fix unintended release (crash) when calling `cast()`
When `transmute_copy()`ing into a local/temporary `InterfaceImpl` (the
generic `IADLXInterface` wrapper to call `.cast()` on), that has drop
semantics and will release the interface one too many times (because it
wasn't created with something like `Clone` which bumps the refcount).

By fixing that the example now no longer SEGFAULTs after querying a
`Gpu1` from `Gpu` and dropping it one too many times at the end of the
loop.